fix(ruff): resolve remaining misc rules down to BLE001-only backlog
MSN-0370 chunk 7/N. Cleared every remaining ruff rule family except
BLE001 in intelligence/ and tools/: ISC004 (8), PIE810 (4), F821 (4),
SIM102 (3), G201 (3), E722 (2), PERF102 (2), N999 (2), FURB192 (2),
C401 (2), SIM113 (1), SIM118 (1), PLW1508 (1), PYI034 (1), PLW0127 (1),
RUF046 (1), RUF015 (1), RUF022 (1), B018 (1) -> 0. Total findings
310 -> 265 (all remaining are BLE001).

Notable non-mechanical fixes:
- tools/integrate_lessons.py had a literal `PYEOF` heredoc terminator
  left in the file after `if __name__ == "__main__": main()` — a real
  leftover artifact from however the file was originally generated, not
  valid Python syntax-adjacent text. Removed (this is what ruff's B018
  useless-expression + F821 undefined-name were actually flagging).
- brief_coherence.py / enrichment_validation.py: bare `except:` ->
  `except (json.JSONDecodeError, ValueError):` (the only exceptions
  json.loads on a string can actually raise), and dropped a genuine
  self-assignment (`cps230 = cps230`) left over from an incomplete edit.
- llm_provider.py / intelligence_store.py: F821 on quoted forward-ref
  type annotations ("Mistral | None", "LLMCallResult | None") referring
  to names only imported lazily inside functions — added
  `if TYPE_CHECKING: import ...` so the annotation resolves for
  static analysis without adding a hard import-time dependency.
- scheduler.py / github_brief_sync.py: `log.error(msg, exc, exc_info=True)`
  -> `log.exception(msg)` (G201/TRY401 — .exception() already captures
  the traceback; passing exc again was redundant, and the unused `exc`
  binding this exposed was then also cleaned up, F841).
- N999 (tools/health-osint/ hyphenated directory name): NOT renamed —
  it's referenced by path across cron/systemd units and other scripts;
  renaming is a real cross-cutting change with production-invocation
  risk, out of scope for a lint-triage mission. Flagged via `# noqa` in
  both __init__.py files with a comment explaining why, not silently
  suppressed without explanation.
- content_intelligence_service.py's known governance gap (flagged in
  chunk 5) is untouched here.

Everything else (ISC004 paren-wrapping, PIE810 startswith-tuple merges,
SIM102 nested-if collapses, FURB192 sorted->min, C401/SIM118/PERF102
comprehension rewrites, RUF046/RUF015/RUF022/PYI034/PLW1508 mechanical
fixes) is a pure syntax rewrite with no behavior change.

Verified: py_compile clean on all changed files; ruff findings in
intelligence/+tools/: 265, all BLE001 (down from 617 at mission start).

SKIP=ruff-check,bandit per .pre-commit-config.yaml's documented
convention — BLE001 findings in these files are the next chunk's scope.

+++ intelligence/analysis/intelligence_analyst.py
@@ -104,7 +104,7 @@ class DualPathScoringResult:
 def _clamp(value: Any) -> int:
     """Coerce a model/heuristic value to an int in [1, 5]; default 3 on garbage."""
     try:
-        v = int(round(float(value)))
+        v = round(float(value))
     except (TypeError, ValueError):
         return 3
     return max(_MIN, min(_MAX, v))
